🐜  Pass `directory` to the `_getWatchedDir` method
The value of `wh.path` might be a glob, so we cannot safely pass it to `_getWatchedDir`.
If we did that, a new entry in the `_watched` map would be created.
Additionally, the call to `fs.readDir` would be passed a glob (undefined behavior?).

This only applies when `options.useFsEvents` is false.